Add client-side and server-side voter eligibility checks
Users who haven’t met the requirements to vote will now see a message and their list of runs on the election page. Additionally, a server-side eligibility check is in place to prevent ballots being saved through other means.

Additionally, the admin page has a list of everyone who is eligible and everyone who’s run in the past year but is not eligible to vote.