Return the “right” name for hashers in the eligibility list
This “fixes” an issue where newboots that had been named at a later run in the year did not have their hash name on the eligibility list list.

This was “fixed” by sorting the runs by their trail number, but trail numbers are sometimes out of order and they can be named after the year ends.